Add ability to print buffers and buffer slots
Examples:

print retrieval
print retrieval.word

Fixes #138
